requirements.txt
streamlit>=1.28.0
google-generativeai>=0.4.0

